Move rating-icons.js to field-rating directory
Move the rating icons module from shared/ to blocks/field-rating/
to keep rating-related code closer to the rating field implementation.
