+++ projects/packages/forms/src/blocks/field-rating/rating-icons.js
